requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
filelock==3.24.2
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.16.0
networkx==3.4.2
numpy==2.2.6
openai==2.21.0
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.27.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
transformers==5.2.0
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0

